[ style ] send code 버튼 컴포넌트 분리
